+++ app.vue
@@ -1,0 +1,14 @@
+<script setup>
+useHead({
+  title: 'Cologne Zoo'
+})
+</script>
+
+<template>
+  <NuxtLayout>
+    <NuxtPage />
+  </NuxtLayout>
+</template>
+
+<style lang='scss'>
+</style>